display square-filled holes differently from floor
This currently does not compile right now because it uses 7 layers, but the maximum is 6. To get around that, I think I will need to either have the game engine changed or expand the 4 HoleBoundaries into all 16 of their permutations.

I tried combining the HoleBoundaries into one collision layer, but that only allows one HoleBoundary to display at a time.