Fix the error messages
I was getting these errors:

```
line 241 : Name "P" already in use.
line 252 : You named an object "<", but this is a keyword. Don't do that!
Errors detected during compilation, the game may not work correctly.
```

I think the “P” error is because objects are now case-insensitive. So I removed the lower-case version, because it is redundant, and I never used that form.

Since ‘<’ is a keyword, I deleted that name for Triangle. I used the ‘\’ form instead, because it looks better in editors than the Unicode one, which sometimes has messed-up spacing.