Split overlapped hole boundary layers
It looks like the engine no longer gives an error when I do this. Enough layers are supported that I can display all four directions of hole boundary when needed.

I regenerated a screenshot so that it reflects this graphical change. The other two screenshots don’t need to be regenerated.