Darken hole outline color for constrast
Hole outlines were the same color as the corners of solid blocks. This looked a little confusing when a hole was next to a solid block (e.g. levels 6, 7, 10).

Now the hole outlines are a bit harder to see and appreciate, especially on dim screens, which is a pity. But I think it’s worth it for holes to look less confusing when next to solid blocks.

I calculated the blended color with http://www.colorhexa.com/color-blend.